[refactor] GGBE-101 change directory structure
